Patch upgrade for Apache Tomcat from 7.0.32 to 7.0.47 to take care of ODL Karaf Northbound issues.
7.0.32 had no support for Cors and hence we added a hack in the past to pull the CorsFilter from Apache Tomcat
project and placed it as a third-party bundle in the controller project.

But the Cors support was added to Tomcat patch version 7.0.42 and hence it is time to upgrade it to the proper
and remove the hack that we added earlier.

Also, the hacked bundle is not loaded in a karaf container (never attaches itself with the Fragment Host).
Debugged the heck of that, but dont know why the hacked bundle wouldnt register.
But since upgrading is cleaner anyways, we are taking that approach.

Tested the Northbound APIs and it is all working fine.

Change-Id: I37bf577320c5257b72e9fbe85313817230507a0c
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>