Update patch set 3
Patch Set 3:

Two concrete requests:

1)  Could we please get a *-source.jar up in the repo for it so that we have the source on record
2)  Could we look at doing this in a way that doesn't slam all the legacy distribution stuff all at once and without warning?  For example, could we not simply use the newer tomcat dependency in the features, without changing everyones legacy distributions over?

I concur we want to get unblocked for karaf.  I'd just like to make sure that we have the source in hand in our repo (item #1) and that folks can transition as they transition to karaf (item #2) so they don't get a sudden surprise in their existing distribution directories as they are in their final push.

Patch-set: 3