Update patch set 1
Patch Set 1:

This works for me - my test:

1) Remove the odlparent /m2 cache
2) Download the odlparent project and this patch
3) run mvn clean install and verify it builds.

I think using the variables here for the repository is OK because the odlparent IS the root pom, so it can resolve the variables correctly. Using variables in the repository section in other projects we found does NOT work because we have the "chicken and the egg" complex, where we need to download the parent from the repository, but we can't contact the repository until we have resolved the variables from the parent.

So, in ODLParent it should be fine to use variables in both the repository and distributionManagement sections. In projects that reference ODLParent as a parent we can only use variables in the distributionManagement section (which is fine because we can actually override the repository section using the "mirror" tag in settings.xml).

If you would prefer leaving the repository section hardcoded that is fine (again we can override it using the mirror tag in settings.xml). But the distributionMangement MUST be variables, otherwise we prevent anyone from uploading artifacts to different repositories during the maven deploy stage.

Patch-set: 1
Reviewer: Gerrit User 1840 <1840@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0