Add pluginManagement section for declared plugin versions
Aside the version declarations, this adds pluginManagement, so we get
warnings when versions get overridden. It also allows downstream users
not to hunt for the proper property name.

This should solve the problem of having duplicate property names, which
we are already getting.

Also flags FIXMEs for convergence of the versions which are clashing.
Hopefully these will be taken care of as the warnings are emitted, as
now users do not have to specify versions at all.

Change-Id: Iae30f52a39e356c1f45b0a189ac75a82e0677695
Signed-off-by: Robert Varga <rovarga@cisco.com>