Update patch set 3
Patch Set 3:

After thinking about this for awhile I came up with 3 solutions (See patchsets 1, 2, and 3.

I decided to settle on Patchset 3 because I think the aggregator root pom does not need to be deployed. Patchset 3 adds configuration to tell Maven to not deploy the root level pom. All other poms are deployed as expected.

Patch-set: 3