Removed <sonar.profile> odlparent/pom.xml.
Daily builds on the master branch have been failing since it is setup
due to the misconfiguration on <sonar.profile> of odlparent/pom.xml.
https://jenkins.opendaylight.org/releng/view/odlparent/job/odlparent-daily-master/

Considering "Sonar way with Findbugs" profile and "Sonar way ODL"
profile, which is the default profile in sonar.opendaylight.org, are
the exactly same, <sonar.profile> is not needed to be set.
https://sonar.opendaylight.org/profiles/compare?id1=1&id2=4

Change-Id: Ida4ab5c7b97c2bab7c51e501241d23bc6ce9f5b8
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>