Update patch set 1
Patch Set 1:

Hi Vaishali,

I'm not sure which version Maven takes if the version is not added.

However, since we don't added the version right now, Maven output the following WARNING messages.
Please check a console output of the odlparent-merge-master job.
To remove the warning messages, I think we need this patch.

https://jenkins.opendaylight.org/releng/view/odlparent/job/odlparent-merge-master/3/consoleFull

[WARNING] Some problems were encountered while building the effective model for org.opendaylight.odlparent:odlparent-aggregator:pom:1.5.0-SNAPSHOT
[WARNING] 'build.plugins.plugin.version' for org.apache.maven.plugins:maven-deploy-plugin is missing. @ line 37, column 21
[WARNING] 
[WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
[WARNING] 
[WARNING] For this reason, future Maven versions might no longer support building such malformed projects.

Patch-set: 1