Add ODL Checkstyle
Adds a checkstyle resource for ODL projects to use that allows them to
take advantage of ODL code style best practices in Java. Details on ODL
Code Style Guidelines can be found on the wiki at
https://wiki.opendaylight.org/view/ODL_Coding_Guidelines

maven-checkstyle-plugin 2.13 does not yet support some of the Google Style
checks which ODL style is heavily based on. Until this is resolved we need
to comment out unsupported checkstyle sections.
https://jira.codehaus.org/browse/MCHECKSTYLE-261

Change-Id: I0097a8fb0ab894d0ee11375e4956f51c30229f24
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>