Update patch set 4
Patch Set 4: Code-Review+1

You are right.

I tested this against the LISPflowmapping project, which defined odlparent as the parent, but that project already used the checkstyle rules from the controller (in warning mode).  With both checkstyle configurations, the build failed, but with the old configuration there are only warnings.

Sorry for the noise.

Patch-set: 4
Label: Code-Review=+1