Bug 2761: Disables external snapshot repositories for feature-test
Pax URL and Karaf by default always search for new version of snapshots
in all snapshots repository, even if that snapshots does not belong to that
repository and maven is invoked even with -nsu (no snapshot update)
or offline mode.

This is also true for OpenDaylight snapshot artefacts - pax url tries
to resolve them from third-party repositories, even if they are not present
there - this increases time which takes for features to install.

For more complex projects this actually means several HTTP GETs for each
snapshot bundle referenced, even if the bundle is already present
in local maven repository.

In order to speed-up installation and remove unnecessary network traffic,
which fails for obvious reasons, external snapshot repositories are
removed.

BGPCEP feture test (all snapshots downloaded, running with mvn clean install -o)
takes 2:16 minutes with this patch. Originally it took 12 minutes and hundreds of
 HTTP GET calls to snapshot repositories were observed.

Change-Id: I939cbfcab86eddc1ba7663e78c37bb57f2d8f0f5
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>