Update patch set 2
Patch Set 2:

Another point, I recall when I tried to enable this by default for all jobs there was push back because some projects thought that the added time to generate these for verify jobs was too long and they wanted it disabled for verify jobs.

Perhaps we can add this into a profile activated via property such as "-Drelease" (I'll need to add this to JJB) but then we can only activate it for merge and daily jobs.

Patch-set: 2