Update patch set 2
Patch Set 2:

You misunderstood me. checkstyle is also odlparent artifact [1] (see the checkstyle directory in odlparent, it has it's own pom.xml which must be bumped for release). It is the module that provides the checkstyle rules for ODL projects. Since this is a module that odlparent releases as an artifact I would say it should be included in the version bump.

 [1] https://git.opendaylight.org/gerrit/gitweb?p=odlparent.git;a=tree

Patch-set: 2