Update patch set 1
Patch Set 1:

> Hi, any reason why this patch was reverted?

It was reverted because it got merged in after Lithium was already built so to ensure that the version bump and tagging matched what was built we had to revert all patches that made it in during the time the final Lithium release was being reviewed.

Sorry I was supposed to unrevert this patch at the same time right after version bumping but forgot to push my unrevert patch online. I see you already submitted the unrevert patch.

Patch-set: 1