Upgrade upstream versions (part 1, Maven)
The versions were checked with versions:display-plugin-updates and
versions:display-dependency-updates.

With the following fixes, all of Beryllium builds, from scratch, with
no external snapshots:
* yangtools: https://git.opendaylight.org/gerrit/#/c/26854/
* controller: https://git.opendaylight.org/gerrit/#/c/26859/
* aaa: https://git.opendaylight.org/gerrit/#/c/26959/
* toolkit: https://git.opendaylight.org/gerrit/#/c/26914/
* dlux: https://git.opendaylight.org/gerrit/#/c/26916/

Specify that Maven 3.1.1 is a prerequisite, as a variable in
odlparent. (The only point of defining the prerequisite is so that the
versions plugin doesn't complain about it. The actual enforcement is
handled by the enforcer plugin.)

All Maven plugins updated to the latest upstream versions, apart from
Karaf (in a later patch in the set).

Change-Id: Ia40a9d1f6428784429ddd8474c39be8984c7da19
Signed-off-by: Stephen Kitt <skitt@redhat.com>