Upgrade upstream versions (part 1, Maven) [DO NOT MERGE]
Remaining questions:
* how do we drive this exactly?
* do we want to upgrade to Spring 4?

The versions were checked with versions:display-plugin-updates and
versions:display-dependency-updates.

Specify that Maven 3.1.1 is a prerequisite, as a variable in
odlparent. (The only point of defining the prerequisite is so that the
versions plugin doesn't complain about it. The actual enforcement is
handled by the enforcer plugin.)

All Maven plugins updated to the latest upstream versions, apart from
Karaf (in a later patch in the set).

Change-Id: Ia40a9d1f6428784429ddd8474c39be8984c7da19
Signed-off-by: Stephen Kitt <skitt@redhat.com>