Remove <repositories> and <pluginRepositories> sections
In January 2015 we made the decision to remove <repositories> and
<pluginRepositories> sections from all ODL projects. This was to meet
the Maven Central requirements that these sections cannot be defined in
your local project poms if you wish to push your artifacts to Maven
Central. Instead we started recommending developers to define these
sections in their local settings.xml file.

https://lists.opendaylight.org/pipermail/discuss/2015-January/004482.html

Change-Id: I4bdf45ed3cbb5e297c8a7ce1165ea63bf917efbb
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>