Update patch set 1
Patch Set 1: Code-Review+1

What's the odlparent rule for committers merging patches?

I typically follow a workflow of not merging my own patches if there are other committers on the project to review so that we get at least 2 sets of eyes.

Please let me know if there's any such rules. Anyway we need this patch merged soon if possible as it fixes sonar coverage report %s for a few projects.

Patch-set: 1