Update patch set 1
Patch Set 1:

Hi Robert, I finally figured out the cause of maven-bundle-plugin forking and executing numerous times. Looks like it's a regression in the version of the maven-bundle-plugin we were using. It is fixed in 2.5.5 and newer. I decided to patch to the latest 3.0.0 and it seems to work with the few projects I tested against. I'm not sure if we want to jump to this version or stick with something in the 2.x stream but at least we know the cause of the unnecessary forking now.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0