Allow child projects to fail on checkstyle errors
When the severity value is set in the odl_checks.xml file it can not
be overwritten in the child projects. Adding a property to the config allows
child projects to override the value. The default is still warning as it is today
and is only overwritten with the changes below in the project pom.

The below is what a child project would add to their pom to override the value:

<plugin>
  <groupId>org.apache.maven.plugins</groupId>
  <artifactId>maven-checkstyle-plugin</artifactId>
  <configuration>
    <propertyExpansion>checkstyle.violationSeverity=error</propertyExpansion>
  </configuration>
</plugin>

Change-Id: Ic271080e4e0b21d49194d66cad72707837cdc16a
Signed-off-by: Sam Hague <shague@redhat.com>