Update patch set 3
Patch Set 3:

Thanks for looking into this!

 > I have noticed the following interesting exception (present also in
 > previous pure-Beryllium run) in https://jenkins.opendaylight.org/releng/job/ovsdb-csit-1node-netvirt-only-beryllium/381/artifact/karaf.log
 > 
 > It probably means that downstream artifacts need to be rebuilt to
 > realize newer Jetty is to be used (meaning integration-patch-test-beryllium
 > is not enough to system test this change).
 > 
 > 2015-10-07 11:41:40,088 | WARN  | qtp1477061003-60 | ServletHandler
 >                   | 169 - org.eclipse.jetty.aggregate.jetty-all-server
 > - 8.1.15.v20140411 | Error for /controller/nb/v2/neutron/networks
 > java.lang.NoClassDefFoundError: javax/validation/ConstraintViolationException

Some of the neutron developers have had this as well, on current master, but we haven't been able to pin it down. It happens on Maven 3.3.3, and goes away with Maven 3.2.5, and once it's gone away with 3.2.5, it works with 3.3.3 too (presumably as long as the ~/.m2/repository is kept)...

It seems a dependency on javax.validation:validation-api:1.1.0.Final is needed somewhere, but I don't like the undeterministic side of things.

Patch-set: 3