Update patch set 13
Patch Set 13: Code-Review+1

This passes integration (discounting the current didm breakage).

Grepping through ODL indicates that the old Spring artifacts are still referred to in snmp4sdn and vtn, but the former overrides spring.version (so is not affected by this patch) and the latter only refers to Spring in manager/it/northbound which is disabled and completely broken (it tries to use ControllerCustomRealm).

Patch-set: 13
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1