Update patch set 3
Patch Set 3:

> Aren't the workspaces preserved?

Not for long enough. Possibly we should make it archive explicitly, as we do for builder-verify job [0].

And I vaguely recall Robert has suggested to package test outputs into jars and upload them to Nexus, similarly to what we do with generated sources (but I may be remembering it wrong).

[0] https://git.opendaylight.org/gerrit/gitweb?p=releng/builder.git;a=blob;f=jjb/releng-jobs.yaml;h=8ce116a7faa95650a916c6ccc93a4bf57932f7b1;hb=109c212e73777c5a03d5b7f97aba5e973b8ef56d#l75

Patch-set: 3