Update patch set 1
Patch Set 1:

> Not sure why distribution-check failed (karaf.log not found) but
 > looks like it successfully built. I think we can merge this one.

I'm getting the failure in integration/distribution with this patch locally too, there's definitely something wrong. I'm working on it, if I can't find the fix soon I'll revert the logback upgrade and upgrade logback in this patch (with the fix for the distribution issue).

Patch-set: 1
Label: Verified=0