Inline versions for org.eclipse.persistence and old property
The "eclipse.persistence.version" property is not used outside of
odlparent;  I pulled autorelease and ensured only odlparent references
this property.  The property was removed and the correct current
version (2.5.2) was inlined instead.

A separate patch will be used to perform the actual upgrade, as that
will be a longer process.

Change-Id: Id39db716ff8f7540ca72670776bbaf017ec98f16
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>