Update patch set 1
Patch Set 1: Code-Review+2

I generally leave the duplicate information, because it avoids errors when modules' parents change (which has happened in the past, e.g. an ovsdb module which is attached to a controller parent POM). This is unlikely to happen in odlparent modules though so we might as well clean things up.

Patch-set: 1
Label: Code-Review=+2