Improved Eclipse IDE experience: m2e's maven-dependency-plugin
It now now longer whines about "Artifact has not been packaged yet. When
used on reactor artifact, unpack should be executed after packaging: see
MDEP-98" nor "Plugin execution not covered by lifecycle
configuration: org.apache.maven.plugins:maven-dependency-plugin
:2.10:unpack (execution: unpack-license, phase: generate-resources)"

Strangely, changing this in odlparent/pom.xml (only) does not work and
leads to an "Error" in Project > Properties > Maven > Lifecycle Mapping,
Copy to Clipboard, and despite M2E offering a Quick Fix to add the
ignore anywhere in the pom.xml hiearchy, it has no effect / causes error
in odlparent - but when added to its parent odlparent-lite then it's OK!
(Or inside the last/leaf pom.xml; but this should be higher up.)

Alternatively it would have been possible to install the m2e connector
for the Maven Dependency Plugin - BUT that only hurts build performance,
and for something as minor as copying a license file around during
build, I don't see the point in-IDE, and so IMHO this one should be
ignore instead of execute.

Change-Id: I3754096209e28fb03ad84db3fe847f0ae776f9d1
Signed-off-by: Michael Vorburger <vorburger@redhat.com>