Cleanup: Checkstyle does NOT require dependency to odl-license
The license check STILL WORKS with this clean-up, because that is done
using odlparent.checkstyle/src/main/resources/LICENSE.txt (which is a
REGEXP for Checkstyle). odlparent.odl-license/src/main/resources/LICENSE
on the other hand is a TXT used only by maven-dependency-plugin
unpack-license, and has nothing to do with Checkstyle. While this
doesn't cause any bugs, it's somewhat confusing, and seemed like a good
idea to clean up for the next person coming along and looking at this.

Originally I hoped that this would resolve an Error that occurs in
Eclipse with the Checkstyle Eclipse plugin, where it gets confused
between the REGEXP license for Checkstyle and the TXT. While this does
not solve that - it still seemed like a good idea to clean up.
(org.apache.commons.beanutils.ConversionException: line 83 in header
specification is not a regular expression at
com.puppycrawl.tools.checkstyle.checks.header.RegexpHeaderCheck.postProcessHeaderLines(RegexpHeaderCheck.java:160)

The removal of pluginManagement in license/pom.xml is unrelated, just
seemed like a good idea too, as that doesn't actually do anything.

Change-Id: I66a0678ebb2fc8d4a4a6f1dd3934979ab0bbd6d8
Signed-off-by: Michael Vorburger <vorburger@redhat.com>