Update patch set 1
Patch Set 1:

Thanh, terribly sorry about this!  I'll use this as an opportunity to try to learn more about whatever it is that "autorelease" does.. PS FYI Just for me focus I will not try again for this one, as it was only a minor suggested clean up - let's forget about it.  My apologies to you again for the overhead this may have caused you!

Patch-set: 1