Clean up more properties
Now that centinel is updated, and since persistence won't release with
Boron (and is still on odlparent 1.6.0-SNAPSHOT), remove the remaining
obsolete dependency properties:
* guava.version
* org.json.version
* jettison.version
* jsr305.api.version
* junit.version
* slf4j.version

To handle the removal of ${guava.version}, hard-code the provided
feature version; this allows us to cleanly provide both versions 18
and 19, in preparation for migration to the latter. Add both versions
to the linked Javadocs.

Change-Id: I767f35ece600b005677299b7bf07af12b4351a4f
Signed-off-by: Stephen Kitt <skitt@redhat.com>