Update patch set 1
Patch Set 1:

Has the community been notified and agreed that we will be modifying our checkstyle rules globally?

I believe last time we discussed this it was agreed we'd follow the Google Style Rules which was what is imported here. I don't think we can start adding new rules in a global way like this without community agreement.

Projects are welcome to override rules locally as they want but things going into odlparent affects all projects so need community input here. When we even introduced this file there was some backlash from the community and it was mandated that it must NOT be configured to fail builds so I want to make sure any changes we make here doesn't come as a surprise to people. The style guide should be updated to reflect changes being made too:

https://wiki.opendaylight.org/view/OpenDaylight_Java_Style_Guide_-_Proposed

Patch-set: 1
Label: Verified=0