Exclude Jackson's org.json bundle
Since Jackson 2.4.0, jackson-datatype-json-org pulls in the Geronimo
bundle of org.json 20131018 (see
https://github.com/FasterXML/jackson-datatype-json-org/issues/6 for
details). This defines a strict dependency on org.json 20090211 (even
though that's not the bundled version), which breaks any resulting
distribution with a later version of org.json.

This patch excludes the bundle. This breaks any project which relies
on jackson-datatype-json-org to provide org.json:json, but those
projects are currently broken anyway.

Change-Id: I03d528539139597dbf3a6c43ed671ebd999f811f
Signed-off-by: Stephen Kitt <skitt@redhat.com>