Update patch set 1
Patch Set 1: Code-Review-1

> Build Failed
 > 
 > https://jenkins.opendaylight.org/releng/job/odlparent-verify-boron/403/
 > : FAILURE

This fails with "Failed during checkstyle configuration: RegexpSinglelineJava is not allowed as a child in Checker" which is a regression.

Apart from that, I'm not sure we want to allow commented out System.out — if it's worth keeping around for debugging purposes, it's worth logging at trace level (perhaps commented out, but as a trace log rather than System.out).

Patch-set: 1
Label: Code-Review=-1