Update patch set 2
Patch Set 2: Code-Review+1

The "RegexpSinglelineJava is not allowed as a child in Checker" problem is fixed now.

Stephen, on further thought and analysis of how this works in Checkstyle, I'm now even more convinced than before that this is The Right Thing(TM) to do ... because: Imagine we'd like to also apply (some) Checkstyle checks to non-*.java file, say src/main/resources (as I'm proposing in https://git.opendaylight.org/gerrit/#/c/40957/), or documentation markup sources - this would make a lot of sense for the FileTabCharacter, "Line has trailing spaces" & "Line has Windows line delimiter" kind of checks - agreed? And Checkstyle can do this.  But what we have with the "Line contains console output" check as-is today is not specific to Java Code, but just a dumb regexp. That means that as-is today, if we ran base Checkstyle against doc resources, if would fail if a say MD contained text like "Thou shalt not use System.out.println" - which surely isn't what we want.  It really does make sense to run this kind of checks only against real code - and thus not anything that is commented out. Un-Commenting it would, of course, make Checkstyle pick it up again (and fail, if the project is so configured). Thus RE +1 from me - hope we can merge this.

Patch-set: 2
Label: Code-Review=+1