Update patch set 2
Patch Set 2:

> > The Optional switch involves a bunch of API changes, so we can't
 > > really drive it from Checkstyle...
 > 
 > Robert Vargas has refused https://git.opendaylight.org/gerrit/#/c/39410/
 > on this ground.  I therefore firmly believe that ODL overall must
 > come to a conclusion about which Optional it wants to use in its
 > code base, then drive changing any existing APIs that use the other
 > one to change, and then merge this.

We've come to a conclusion — java.util.Optional is the way forward (which is why it's used for new code where Guava's version isn't made necessary by existing APIs). But changing all the uses of Guava's Optional to java.util.Optional isn't going to happen overnight, it's a Carbon target at the earliest.

Patch-set: 2