Relax Checkstyle for JavaDoc, not mandatory for annotations
Because in at least one instance a project (ovsdb/library;
https://git.opendaylight.org/gerrit/#/c/40967/) has adopted Checkstyle
99% but then disabled enforcement because of lack of JavaDoc in a few
annotation types.  This is the wrong way to go about this IMHO; I'd
rather we relax this arbitrary requirement (that even annotations do not
have to have JavaDoc, like everything else), and be able to enforce
Checkstyle in more projects.

Change-Id: I42d1c81aa5ff511c272d39752340dddc5e3b0636
Signed-off-by: Michael Vorburger <vorburger@redhat.com>