Introduce PMD Copy/Paste Detector (CPD)
Like Checkstyle, this by default only reports to the console during
(local..) builds, but does not enforce it - builds will NOT fail with
this. Then projects ready to "opt-in" for CPD enforcement can use:

  <properties>
    <pmd.cpd.fail>false</pmd.cpd.fail>
  </properties>

BTW: The "minimumTokens" threshold may take some tuning - let's start
with these approx. 10 line rule, and then adjust it as we see how this
plays out in practice on a number of projects.

Change-Id: If862b8a16db5eed391bb85165987a4302ae44540
Signed-off-by: Michael Vorburger <vorburger@redhat.com>