Intro. Xtend dependencyManagement and Maven plugin
Xtend <http://xtend-lang.org> is already used in ODL's MDSAL project,
and I'd like to start using it in other projects as well (for writing
test code, for now).  In order to keep the version of Xtend used
uniform, and to keep child pom.xml shorter re. the xtend-maven-plugin
configuration, it would thus make sense to have this in the ODL parent.

Change-Id: I8356ef07772d3d487b688b5f6789c23a9b698bd2
Signed-off-by: Michael Vorburger <vorburger@redhat.com>