Introduce PMD Copy/Paste Detector (CPD)
Take II - is using <excludeRoots><excludeRoot> better? (We tried this
before and had to revert, despite it passing on Jenkins; very strange.)

Like Checkstyle, this by default only reports to the console during
(local..) builds, but does not enforce it - builds will NOT fail with
this. Then projects ready to "opt-in" for CPD enforcement can use:
<properties> <pmd.cpd.fail>false</pmd.cpd.fail> </properties>

Change-Id: I01de6ce01fe5f7020aae47ee2e9325cdc14e00e4
Signed-off-by: Michael Vorburger <vorburger@redhat.com>