Checkstyle IllegalThrows with ignoreOverriddenMethods
The default is true, and for good reason; it was stupid of me to
originally have used ignoreOverriddenMethods = false.

This is because when you @Override a method, you cannot do anything
anymore about its throws clause - you have to comply with what you are
overriding.

Based on doing some real world Checkstyle clean up work and noticing
that this was wrong.

Change-Id: Ifbfd677074d7b84cbea52f5ca520791d648def75
Signed-off-by: Michael Vorburger <vorburger@redhat.com>