Activate PMD CPD by default
The new Copy-Paste detector which I recently introduced in
c5cfbc434448b646fd93a89e0928111e2a4d500c and announced on
https://lists.opendaylight.org/pipermail/odlparent-dev/2016-July/000443.html
would probably be of more value if it was actually on by default, and
not just dormant in an pluginManagement doing nothing? ;-)  Note again
that this will never break a build - just log; unless enforcement is
enabled (same model as for Checkstyle).

Change-Id: I07b90dfb5a9c3f724eba77d2b87e08c25699c8b6
Signed-off-by: Michael Vorburger <vorburger@redhat.com>