Update patch set 5
Patch Set 5:

> > Why shouldn't we upgrade?
 > 
 > You will not be able to start karaf using an absolute path, see
 > https://issues.apache.org/jira/browse/KARAF-4599
 > 
 > But maybe it isn't that important. I just wanted to warn about
 > this.
 > 
 > karaf script will work fine
 > start/stop/status/... won't support being called using an absolute
 > path.
 > 
 > So for sure, devs aren't going to be too impacted because I believe
 > the karaf script is the most used.

Can't we just apply the fix from upstream on top of our distribution?  I'd elect that if everything else works.  I have been building autorelease on a node all day with this change... I am hoping we can get it in for Boron :).  Open to differing opinions of course...

Patch-set: 5