Update patch set 2
Patch Set 2:

(1 comment)

> In addition, before adding a dependency I like to see an example
 > use-case, unless it's immediately obvious that it's useful. So here
 > I'd like to see a patch-set making use of Dagger in some project
 > before we add it to odlparent...

Certainly.  Coming up (possibly week after) next!

 > Additionally, I'd like to keep things consistent in ODL: we have
 > CSS (for some form of DI), we're moving to Blueprint, so

yeah no Blueprint is OSGi.  This is for DI within a bundle, starting with tests.  Complements blueprint, and may even be able to integrate with it in the future.  Not competing.

What is CSS for some form of DI?

Patch-set: 2
Label: Verified=0