Update patch set 1
Patch Set 1: Code-Review-1

Doesn't this result in a Windows 10-specific build? I.e., a Karaf distribution which only works on Windows (any version I guess...).

Does Karaf work if you add

    -Dorg.osgi.framework.os.name=Win32

to karaf.bat, in the line used to start the JVM at the end of the file? (I don't have a Windows system to test on.)

Patch-set: 1
Label: Code-Review=-1