Update patch set 2
Patch Set 2:

I would like to keep it as a placeholder. Once PMD grows the ability to wildcard the path, we really want to make it a project-wide setting, because matching execution/goal downstream is pain.

It also affects things which use other plugins -- antlr in yangtools and xtend in mdsal -- a blanked config to not consider target/generated-sources is the way to go, at least IMHO.

Patch-set: 2