Update patch set 2
Patch Set 2:

On a more practical level, if we don't add the headers at creation time, when people do add more useful content to the package-info.java (hope springs eternal) they're unlikely to think of adding the header too!

This is the kind of aspect which really needs agreement from the LF before we can add an exception (probably Phil Robb).

Patch-set: 2