Update patch set 5
Patch Set 5: Code-Review+1

I think this is now ready for discussion, along with https://git.opendaylight.org/gerrit/38638 and https://git.opendaylight.org/gerrit/39468. With Karaf 3.0.8 the resulting features.xml are identical to the original static files.

This gives us features which are also Maven artifacts, and can be depended on instead of picking and mixing upstream bundles. It also ensures that features are built correctly and don't introduce dependency loops. I'm hoping it will also eventually allow us to statically verify features.xml before running SingleFeatureTest, which will save time on breaking builds.

The main weakness currently is that upstream repositories still have to be specified manually; that is fixed in Karaf 4.1. The impact here is that we need to carry small feature.xml files with overrides, which seems minor to me; we'd perhaps need to adapt the version bumping scripts.

Version ranges aren't supported in Karaf 3.0.8 but are supported in 4.1.

I don't particularly like the approach I've taken in odlparent for upstream features, but it's the best compromise I could come up with — it doesn't break current downstream, and it allows version bumps to be handled reasonably gracefully.

Patch-set: 5
Label: Code-Review=+1