Update patch set 2
Patch Set 2:

> > Well, I think it is fair to require package-info.java to contain
 > > javadocs. At least it should explain what should and should not
 > go
 > > into a particular package...
 > 
 > +1!

Except that this isn't the question here.   We have earlier agreed to, and aligned our Checkstyle rules accordingly, not enforce JavaDoc presence, just validate its content if it's present. However, we are still enforcing every package-info.java (of which we have almost none..) to have a JavaDoc.  This change proposes to make our JavaDoc enforcement rules (which is to not enforce JavaDoc, but to check it for validity if its present) "consistent" for package-info.java with all other *.java.  This seems logical.  If you -1 or (simply ignore) this proposed change, it will probably not lead to people creating useful JavaDoc for existing packages in package-info.java when adding it for @NonNullByDefault, but just to bogus empty JavaDoc to shut up Checkstyle.

Patch-set: 2