Update patch set 1
Patch Set 1:

I will point out that I have made this change in our downstream distribution;  we utilize BouncyCastle as the default JCE provider.  Without this change, our CI literally freezes due to lack of entropy (this one was tough to figure out... why are all of our tests sporadically failing)?  With this change in place, we don't seem to have the mysterious hangs.

Patch-set: 1