Update patch set 1
Patch Set 1:

Vratko, thank you for adding me as a reviewer to this proposed change. Unfortunately my bash experience is quite limited, and I therefore feel unqualified to comment +1/-1 re. removing braces in bash expressions (I do not know what they are good / bad for). Stephen may have more to say?  As a more general comment from me, I would be hesitant to change this scripts which we've effectively forked from upstream Karaf, and would suggest that we should keep them as close as possible to the original Karaf. Anything we change for good reason, we should try to (also) contribute upstream to Karaf, IMHO.

Patch-set: 1