Add configuration and execution to FindBugs plugin
Added the appropriate configuration and execution to the FindBugs plugin
so it can run during compile phase and report violations. failOnError is
set to false by default so it doesn't fail the build. Projects can override
this setting if they want enforcement.

Also added a findbugs project (similar to checkstyle) that defines the rule
filters to exclude. I added a couple to exclude b/c they're of low priority
and are common patterns used in ODL.

Change-Id: Ie01507a23990280d4bee8c2aefdc6c2293a3c6db
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>
Signed-off-by: Stephen Kitt <skitt@redhat.com>
Signed-off-by: Michael Vorburger <vorburger@redhat.com>