Update patch set 13
Patch Set 13:

> (1 comment)
 > 
 > Still +1.
 > 
 > The implied version cleanup can be done in a separate change, as I
 > expect there are many places outside this Change which are also
 > specifying plugin versions needlessly.

Yes, I understand the problem;  we discussed the need for using older versions in a separate patch [0] already. What I meant is more general across all ODL projects; it would be nice to find a way to upgrade artifacts that are unable to use the standard version, 3.0.8, so we can universally use the latest 3.0.X version.  That said, we may be moving to either 4.0.X or 4.1.X in Carbon (I hope!).

Still +1 from me too, as the extraneous version tag doesn't do any harm and can be removed in a later patch. 

[0] https://git.opendaylight.org/gerrit/#/c/45646/

Patch-set: 13