Update patch set 12
Patch Set 12: Code-Review+1

After a bit pulling this down and asking Luis for a bit of help, it was discovered that we still need setenv for Xmx (max heap size).  I have un-unforked the setenv script in this update of the patch, which passed last CI round.  I had to update the commit message as it was no longer appropriate, but this should be good to go after we get a clean CI run.

Patch-set: 12
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1