Update patch set 3
Patch Set 3:

(1 comment)

This makes me a bit nervous;  I really hope that projects don't pull in the karaf4 artifacts before we are ready to make the switch from karaf 3->4.  There is no real way to enforce it at this point since it would appear the plan is to have the two live simultaneously.

Patch-set: 3
Label: Verified=0