Update patch set 3
Patch Set 3:

> This makes me a bit nervous;  I really hope that projects don't
 > pull in the karaf4 artifacts before we are ready to make the switch
 > from karaf 3->4.  There is no real way to enforce it at this point
 > since it would appear the plan is to have the two live
 > simultaneously.

In the current state, they'll just break violently. Once the karaf4 artifacts work, projects should work with either set of artifacts transparently; the deciding factor will be integration/distribution. I don't think any of the karaf-parent children end up in the distribution, apart from integration/distribution's own distribution-karaf project...

Anyway, once karaf4 works the idea is to switch to it rapidly. I'm building this as a separate set of artifacts based on Vratko's idea, so that if necessary (i.e. the amount of work to be done means it's useful for things to be merged progressively) we can merge patches without breaking anything. It's looking like this won't be necessary after all, but it's still useful for me since I can build Karaf 3 and 4 distributions from the same Maven project.

Patch-set: 3